fix a performance bug
when `vvm.install_vyper()` was called with `show_progress=True`, it was
chunking into 1024 byte chunks. updating the progress bar with tqdm
every 1kb was slowing down the download by about 25x(!).

this commit changes the chunk_size from 1024 to `None`, which, according
to the `requests` docs, reads data in whatever size the chunks are
received from the underlying stream.

on my laptop, this reduces download time from ~50s to ~2s.